requirements.txt
mcp-agent
python-dotenv
google-generativeai
